Feat: add deletePost now cheks if user owns post
deletePost checks if the logged user owns the post before deleting it

+++ src/main/java/com/example/blog/service/PostService.java
@@ -91,10 +91,16 @@ public class PostService {
     }
 
     public void deletePost(Long postId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User currentUser = (User) authentication.getPrincipal();
+
         Optional<Post> postOptional = this.postRepository.findById(postId);
 
         if(postOptional.isEmpty())
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Post not found.");
+
+        if(currentUser.getId() != postOptional.get().getId())
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
 
         this.postRepository.delete(postOptional.get());
     }
